test: consolidate battleCLI Escape tests
